+++ GMI_GridComp/GmiChemistry/AerosolDust/GmiAerDustODSA_mod.F90
@@ -561,7 +561,7 @@
 ! !LOCAL VARIABLES:
       REAL*8       :: MSDENS(NSADdust)
 
-      integer      :: i, j , k, l, n, r
+      integer      :: i, j, k, l, n, r
 !
 ! !REVISION HISTORY:
 !   February2005, Jules Kouatchou (Jules.Kouatchou.1@gsfc.nasa.gov)